修改bug

+++ MidnightArtifact/app/src/main/java/com/atman/wysq/model/response/GetMyUserIndexModel.java
@@ -178,7 +178,7 @@ public class GetMyUserIndexModel {
 
             private UserExtBean userExt;
             private int follow_count;
-            private int followed_count;
+            private long followed_count;
             private int favorited_count;
             private int comment_count;
             private int vipLevel;
@@ -234,11 +234,11 @@ public class GetMyUserIndexModel {
                 this.follow_count = follow_count;
             }
 
-            public int getFollowed_count() {
+            public long getFollowed_count() {
                 return followed_count;
             }
 
-            public void setFollowed_count(int followed_count) {
+            public void setFollowed_count(long followed_count) {
                 this.followed_count = followed_count;
             }
