+++ MidnightArtifact/app/src/main/java/com/atman/wysq/model/response/GetUserInfoModel.java
@@ -74,7 +74,7 @@ public class GetUserInfoModel {
 
         private UserExtEntity userExt;
         private int follow_count;
-        private int followed_count;
+        private long followed_count;
         private int favorited_count;
         private int comment_count;
         private String achieve_level_pic;
@@ -120,11 +120,11 @@ public class GetUserInfoModel {
             this.follow_count = follow_count;
         }
 
-        public int getFollowed_count() {
+        public long getFollowed_count() {
             return followed_count;
         }
 
-        public void setFollowed_count(int followed_count) {
+        public void setFollowed_count(long followed_count) {
             this.followed_count = followed_count;
         }
